Refactor and optimize series handling with enhanced downsampling and robust fanout (#21)
Major refactor:  Enhanced memory efficiency, performance, and test coverage

This PR incorporates extensive refactoring, optimizations, and architectural
improvements to series handling, downsampling, and fanout operations for
valkey-timeseries.

## Core Architecture Changes

### Threading & Concurrency
- Introduced new low-latency parallel library for improved concurrent operations
- Added a thread pool to push expensive operations off the main thread

### Memory Optimization
- Implemented string interning for all series labels to reduce memory footprint
- String buffer pooling to reduce allocator pressure

### Compaction & Downsampling
- Streamlined and parallelized downsampling pipeline

### A More Robust Fanout Architecture
- Redesigned fanout implementation using Valkey's cluster bus
- Integrated Protocol Buffers for efficient inter-node messaging and forward compatibility

## Data Layer Improvements

### Range Iteration & Queries
- Enhanced iterator pipeline supporting traversal, filtering, aggregation, and grouping

### Series Management
- Improved index management and expanded querying capabilities (e.g. filter by series with data in a range)

## Command Interface

### Command Parsing
- Refactored command parsing for better maintainability
- Standardized error handling across all commands
- Improved response formatting consistency

## Testing & Quality

### Test Coverage
- Added 500+ unit tests covering core functionality
- Implemented 540+ integration tests for end-to-end scenarios
- Enhanced test coverage across all major components
- Improved test reliability and determinism

## Impact Summary

- **262 files changed**: 41,093 additions, 10,128 deletions
- **703 commits** consolidated from the unstable branch
- Improvements in write throughput and query latency
- Enhanced reliability for fanout operations
- Improved memory efficiency through string interning
- Better test coverage and code maintainability

Closes #21

+++ src/tests/generators/rand.rs
@@ -133,7 +133,6 @@ pub fn generate_series_data(options: &DataGenerator) -> Vec<Sample> {
         let end = if let Some(end) = options.end {
             end
         } else {
-            // todo: automatically choose based on number of samples
             options.start + (options.samples * 1000 * 60) as i64
         };
         (end - options.start) / options.samples as i64
@@ -150,12 +149,12 @@ pub fn generate_series_data(options: &DataGenerator) -> Vec<Sample> {
 
     if let Some(significant_digits) = options.significant_digits {
         for v in values.iter_mut() {
-            let rounded = round_to_sig_figs(*v, significant_digits as i32);
+            let rounded = round_to_sig_figs(*v, significant_digits as u8);
             *v = rounded;
         }
     } else if let Some(decimal_digits) = options.decimal_digits {
         for v in values.iter_mut() {
-            let rounded = round_to_decimal_digits(*v, decimal_digits as i32);
+            let rounded = round_to_decimal_digits(*v, decimal_digits as u8);
             *v = rounded;
         }
     }
